Implemented basic pulse debounce
While our system does have a hardware debouncer for ~1ms glitches, the
nature of the Wall-O-Matic mechanism makes longer glitches possible.
This is especially true if the mechanism is old and no longer making
perfect contact.

To solve this, the ISR now merges any pulse with a sub-10ms gap from the
previous pulse. That seemed like a good ballpark, since testing has
shown glitches as high as 6ms and a real pulse gap is at least 20ms.